Changed how resetting UserQuestionHistory is done. now removes only questions that have corespond to the query instead of all users data. added some indexes and fixed tests for UserQuestionHistoryService.cs

+++ Api/Services/UserQuestionHistoryService.cs
@@ -1,17 +1,28 @@
 using Microsoft.EntityFrameworkCore;
 using QuestionaireApi.Interfaces;
 using QuestionaireApi.Models.Database;
+using SharedStandard.Models;
 
 namespace QuestionaireApi.Services;
 
 public class UserQuestionHistoryService(QuestionaireDbContext context) : IUserQuestionHistoryService
 {
-    public async Task ResetUserQuestionHistory(string userId)
+    public async Task ResetUserQuestionHistoryForCriteria(string userId, IEnumerable<int> categoryIds, IEnumerable<Difficulty> difficulties)
     {
         try
         {
+            List<int> questionIdsMatchingCriteria = await context.Questions
+                .Where(q => q.IsApproved == true && q.IsDeleted == false)
+                .Where(q => difficulties.Contains(q.Difficulty))
+                .Where(q => q.QuestionCategories.Any(qc => categoryIds.Contains(qc.CategoryId)))
+                .Select(q => q.Id)
+                .Distinct()
+                .ToListAsync();
+
+            if (!questionIdsMatchingCriteria.Any()) return;
+            
             await context.UserQuestionHistory
-                .Where(h => h.UserId == userId)
+                .Where(h => h.UserId == userId && questionIdsMatchingCriteria.Contains(h.QuestionId))
                 .ExecuteDeleteAsync();
         }
         catch (Exception ex)
@@ -31,7 +42,7 @@ public class UserQuestionHistoryService(QuestionaireDbContext context) : IUserQu
                 QuestionId = q,
                 RoundNumber = 1
             }));
-
+            
             await context.SaveChangesAsync();
         }
         catch (Exception ex)
